# Translate attributes in `vec_normalize_encoding()`