--- a/tests/testthat/test-translate.R
+++ b/tests/testthat/test-translate.R
@@ -95,24 +95,29 @@ test_that("translation treats data frames elements of lists as lists (#1233)", {
   expect_equal_encoding(result_field, expect_field)
 })
 
-# FIXME: Should we translate attributes to fix this? Can it be done efficiently?
-test_that("attributes are currently not translated", {
+test_that("attributes are translated", {
   utf8 <- encodings()$utf8
   latin1 <- encodings()$latin1
 
   a <- structure(1, enc = utf8)
   b <- structure(1, enc = latin1)
-  x <- list(a, b)
+  c <- structure(1, enc = list(latin1))
+  x <- list(a, b, c)
 
   result <- vec_normalize_encoding(x)
 
   a_enc <- attr(result[[1]], "enc")
   b_enc <- attr(result[[2]], "enc")
+  c_enc <- attr(result[[3]], "enc")[[1]]
 
-  # Ideally both would be utf8
   expect_equal_encoding(a_enc, utf8)
-  expect_equal_encoding(b_enc, latin1)
+  expect_equal_encoding(b_enc, utf8)
+  expect_equal_encoding(c_enc, utf8)
 
-  # Ideally the list elements are considered duplicates
-  expect_identical(vec_unique(x), x)
+  expect <- list(
+    structure(1, enc = utf8),
+    structure(1, enc = list(utf8))
+  )
+
+  expect_identical(vec_unique(x), expect)
 })