Closes #1258 

After much head-banging, I think I've finally gotten the recursive pattern correct in such a way that we can elegantly translate attributes recursively. It required rewriting it again, but I think it is overall even more elegant than before.

This is a bit slower than what we used to do (not translate attributes), but it really is only noticeable if you have deeply nested lists. The most common place this could arise is with multiple list columns, as is often done in tidymodels.

That said, translating the attributes here means that we can rip out the character translation parts of `equal.h` and `compare.c`, which I'm hoping will recover a bit of the lost performance.

Overall, I think it is worth it to put the translation code in one place, so we don't have to worry about it in other vctrs functions.

Here are a few examples, side by side with `vec_unique()`, which internally calls `vec_normalize_encoding()`, so you can get a feel for its impact.

Seems to have no impact on a sampled version of `flights`, which I consider to be fairly realistic.

``` r
library(nycflights13)
library(tidymodels)
library(vctrs)

vec_normalize_encoding <- vctrs:::vec_normalize_encoding

# sample the first 5 rows ~400k times
# 19 cols, 1 has attributes
flights <- vec_slice(flights, sample(1:5, nrow(flights), replace = TRUE))

# A fairly realistic dataset
bench::mark(
  vec_unique(flights),
  vec_normalize_encoding(flights), 
  iterations = 100,
  check = FALSE
)

# before
#> # A tibble: 2 x 6
#>   expression                           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(flights)               38.7ms  41.33ms      24.0    3.31MB     1.53
#> 2 vec_normalize_encoding(flights)    2.2ms   2.39ms     416.         0B     0

# after
#> # A tibble: 2 x 6
#>   expression                           min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                      <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(flights)              36.03ms  38.84ms      25.8    3.31MB     1.65
#> 2 vec_normalize_encoding(flights)   1.88ms   2.17ms     472.         0B     0
```

Does have some impact on a deeply nested tibble that occurs naturally in tidymodels, but it isn't too bad.

```r
resamples <- bootstraps(mtcars, times = 100)

spline_rec <- recipe(mpg ~ ., data = mtcars) %>%
  step_ns(disp) %>%
  step_ns(wt)

lin_mod <- linear_reg() %>%
  set_engine("lm")

spline_res <- fit_resamples(lin_mod, spline_rec, resamples)

# 100 rows, some list columns that have tibbles in them
spline_res
#> # Resampling results
#> # Bootstrap sampling 
#> # A tibble: 100 x 4
#>    splits          id           .metrics         .notes          
#>    <list>          <chr>        <list>           <list>          
#>  1 <split [32/13]> Bootstrap001 <tibble [2 × 3]> <tibble [0 × 1]>
#>  2 <split [32/9]>  Bootstrap002 <tibble [2 × 3]> <tibble [0 × 1]>
#>  3 <split [32/13]> Bootstrap003 <tibble [2 × 3]> <tibble [0 × 1]>
#>  4 <split [32/13]> Bootstrap004 <tibble [2 × 3]> <tibble [0 × 1]>
#>  5 <split [32/10]> Bootstrap005 <tibble [2 × 3]> <tibble [0 × 1]>
#>  6 <split [32/10]> Bootstrap006 <tibble [2 × 3]> <tibble [0 × 1]>
#>  7 <split [32/11]> Bootstrap007 <tibble [2 × 3]> <tibble [0 × 1]>
#>  8 <split [32/11]> Bootstrap008 <tibble [2 × 3]> <tibble [0 × 1]>
#>  9 <split [32/12]> Bootstrap009 <tibble [2 × 3]> <tibble [0 × 1]>
#> 10 <split [32/9]>  Bootstrap010 <tibble [2 × 3]> <tibble [0 × 1]>
#> # … with 90 more rows

bench::mark(
  vec_unique(spline_res),
  vec_normalize_encoding(spline_res), 
  iterations = 100,
  check = FALSE
)

# before
#> # A tibble: 2 x 6
#>   expression                             min  median `itr/sec` mem_alloc
#>   <bch:expr>                         <bch:t> <bch:t>     <dbl> <bch:byt>
#> 1 vec_unique(spline_res)             777.3µs 852.6µs     1133.     103KB
#> 2 vec_normalize_encoding(spline_res)  11.7µs  11.8µs    76472.        0B
#> # … with 1 more variable: `gc/sec` <dbl>

# after
#> # A tibble: 2 x 6
#>   expression                             min  median `itr/sec` mem_alloc
#>   <bch:expr>                         <bch:t> <bch:t>     <dbl> <bch:byt>
#> 1 vec_unique(spline_res)             823.8µs 886.9µs     1116.     103KB
#> 2 vec_normalize_encoding(spline_res)  63.1µs  66.2µs    14950.        0B
#> # … with 1 more variable: `gc/sec` <dbl>
```