@DavisVaughan 2020-09-16
One more example that is a bit simpler, and really shows that there is a performance impact here with lists that we might have to just accept.

```r
library(vctrs)

vec_normalize_encoding <- vctrs:::vec_normalize_encoding

x <- rep_len(list(1), 1e6)

bench::mark(
  vec_unique(x),
  vec_normalize_encoding(x),
  check = FALSE,
  iterations = 50
)

#before 
#> # A tibble: 2 x 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(x)                 23ms  24.49ms      39.9    11.8MB     18.8
#> 2 vec_normalize_encoding(x)    1.6ms   1.73ms     548.         0B      0

# after
#> # A tibble: 2 x 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_unique(x)              29.38ms  31.13ms      30.9    11.8MB     14.6
#> 2 vec_normalize_encoding(x)   7.44ms   8.23ms     121.         0B      0
```