@DavisVaughan 2020-09-23
@lionel- I _think_ I have this correct, but could use a second set of eyes.

This tries to be as efficient as possible by copying the pairlist up to the element that needs translation, and then it uses the tail of the rest of the pairlist.

If a second element need translation, then it only copies nodes after the node of the first element that needed translation. So it reuses a "head" and "tail" and only makes a copy of the "middle".