@DavisVaughan 2020-09-23
I'm not sure if this protect is correct?
https://github.com/r-lib/vctrs/blob/f71c8e0b412bf344ada754b4bfa5fdf03a86cab8/src/translate.c#L221

I made an educated guess based on the fact that `val` gets protected here
https://github.com/wch/r-source/blob/bac4cd3013ead1379e20127d056ee036278b47ff/src/main/duplicate.c#L266