@lionel- 2020-09-29
I think this implementation is a bit too complicated. Ideally we'd have some sort of `node_clone_until()` function that clones the top of a pairlist. It would return the new head and the new tail in an output parameter.

In this case though, since attribute pairlists tend to be short, we could just use shallow duplication of the whole pairlist with a boolean flag to keep track of whether we own the pairlist or not.